BAEL-5699
Removed unused class FifoFixedSizeQueueRemoveOldestInOffer.java
Removed unused UnitTest for the removed class
Added some tests to check count attribute
